debug types

+++ app/models/story.rb
@@ -3,6 +3,6 @@ class Story
   field :user, type: String
   field :story, type: String
   field :about_who, type: String
-  field :lat, type: Double
-  field :lon, type: Double
+  field :lat, type: Integer
+  field :lon, type: Integer
 end
